# auth-service/openapi.yaml
openapi: 3.1.0
info:
  title: Auth Service Management API
  version: 1.0.0
  description: Management API for the Goaldone Auth Service. Provides endpoints for user, organization, and invitation management. All endpoints require OAuth2 Bearer token authentication.
  contact:
    name: Goaldone Team
    url: https://goaldone.de
  license:
    name: Proprietary
servers:
  - url: http://localhost:9000
    description: Local development server
  - url: https://auth.goaldone.de
    description: Production server
security:
  - Bearer: []
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained from the OAuth2 authorization server
  schemas:
...
    CompanyRequest:
      type: object
      title: CompanyRequest
      description: Request body for creating or updating an organization
      properties:
        name:
          type: string
          description: Organization name
          example: Acme Corporation
        slug:
          type: string
          description: Organization unique slug identifier (lowercase alphanumeric and hyphens)
          example: acme-corp
          pattern: ^[a-z0-9-]+$
      required:
        - name
        - slug
    CompanyResponse:
      type: object
      title: CompanyResponse
      description: Response containing organization information
      properties:
        id:
          type: string
          format: uuid
          description: Unique organization identifier
          example: 550e8400-e29b-41d4-a716-446655440002
        name:
          type: string
          description: Organization name
          example: Acme Corporation
        slug:
          type: string
          description: Organization unique slug identifier
          example: acme-corp
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the organization was created
...
    ProblemDetail:
      type: object
      description: HTTP problem detail response (RFC 7807)
      properties:
        type:
          type: string
          description: A URI reference that identifies the problem type
        title:
          type: string
          description: A short, human-readable summary of the problem
        status:
          type: integer
          description: The HTTP status code
        detail:
          type: string
          description: A human-readable explanation of the problem
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence
paths:
...
  /api/v1/organizations:
    post:
      tags:
        - Organizations
      summary: Create a new organization
      description: Creates a new organization in the system
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRequest'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
        '409':
          description: Organization already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/organizations/{id}:
    get:
      tags:
        - Organizations
      summary: Get organization by ID
      description: Retrieves an organization by its UUID
      operationId: getOrganizationById
      parameters:
        - name: id
          in: path
          required: true
          description: Organization UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Organization found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '401':
          description: Unauthorized
        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    put:
      tags:
        - Organizations
      summary: Update organization
      description: Updates an existing organization's information
      operationId: updateOrganization
      parameters:
        - name: id
          in: path
          required: true
          description: Organization UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRequest'
      responses:
        '200':
          description: Organization updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/organizations/search:
    get:
      tags:
        - Organizations
      summary: Search organization by slug
      description: Retrieves an organization by its slug (unique identifier)
      operationId: getOrganizationBySlug
      parameters:
        - name: slug
          in: query
          required: true
          description: Organization slug
          schema:
            type: string
      responses:
        '200':
          description: Organization found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '401':
          description: Unauthorized
        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /api/v1/users/{userId}/memberships/{companyId}:
    delete:
      tags:
        - Memberships
      summary: Remove user from organization
      description: Deletes a user's membership in an organization. Cannot remove the last administrator.
      operationId: deleteMembership
      parameters:
        - name: userId
          in: path
          required: true
          description: User UUID
          schema:
            type: string
            format: uuid
        - name: companyId
          in: path
          required: true
          description: Organization UUID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Membership removed successfully
        '401':
          description: Unauthorized
        '404':
          description: User or organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Cannot remove the last administrator from the organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    patch:
      tags:
        - Memberships
      summary: Update user role in organization
      description: Changes a user's role in an organization. Cannot demote the last administrator.
      operationId: updateMembershipRole
      parameters:
        - name: userId
          in: path
          required: true
          description: User UUID
          schema:
            type: string
            format: uuid
        - name: companyId
          in: path
          required: true
          description: Organization UUID
          schema:
            type: string
            format: uuid
        - name: newRole
          in: query
          required: true
          description: New role (COMPANY_ADMIN or USER)
          schema:
            type: string
            enum: [COMPANY_ADMIN, USER]
      responses:
        '200':
          description: Membership role updated successfully
        '401':
          description: Unauthorized
        '404':
          description: User or organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Cannot demote the last administrator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...